1.8.7-2: Detect Vuzes from /opt

+++ PKGBUILD
@@ -14,8 +14,10 @@ noextract=("${_name}_${pkgver}.jar")
 md5sums=('6441e5ddceb8f4a17116bde6ba693c84')
 
 package () {
-  for i in /usr/lib/vuze*; do
-    install -Dm644 ${_name}_${pkgver}.jar \
-                   "${pkgdir}"/${i}/plugins/${_name}/${_name}_${pkgver}.jar
+  for i in /usr/lib/vuze* /opt/vuze*; do
+    if [[ -f ${i}/Azureus2.jar ]]; then
+      install -Dm644 ${_name}_${pkgver}.jar \
+                     "${pkgdir}"/${i}/plugins/${_name}/${_name}_${pkgver}.jar
+    fi
   done
 }
